requirements.txt
async-timeout==5.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.2
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.2.0
requests==2.32.4
seaborn==0.13.2
six==1.17.0
spotipy==2.25.1
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
